+++ source/extensions/common/dynamic_forward_proxy/dns_cache_impl.cc
@@ -2,6 +2,7 @@
 
 #include "envoy/extensions/common/dynamic_forward_proxy/v3/dns_cache.pb.h"
 
+#include "common/config/utility.h"
 #include "common/http/utility.h"
 #include "common/network/utility.h"
 
@@ -15,7 +16,7 @@ namespace DynamicForwardProxy {
 
 DnsCacheImpl::DnsCacheImpl(
     Event::Dispatcher& main_thread_dispatcher, ThreadLocal::SlotAllocator& tls,
-    Stats::Scope& root_scope,
+    Runtime::RandomGenerator& random, Stats::Scope& root_scope,
     const envoy::extensions::common::dynamic_forward_proxy::v3::DnsCacheConfig& config)
     : main_thread_dispatcher_(main_thread_dispatcher),
       dns_lookup_family_(Upstream::getDnsLookupFamilyFromEnum(config.dns_lookup_family())),
@@ -23,6 +24,10 @@ DnsCacheImpl::DnsCacheImpl(
       scope_(root_scope.createScope(fmt::format("dns_cache.{}.", config.name()))),
       stats_{ALL_DNS_CACHE_STATS(POOL_COUNTER(*scope_), POOL_GAUGE(*scope_))},
       refresh_interval_(PROTOBUF_GET_MS_OR_DEFAULT(config, dns_refresh_rate, 60000)),
+      failure_backoff_strategy_(
+          Config::Utility::prepareDnsRefreshStrategy<
+              envoy::extensions::common::dynamic_forward_proxy::v3::DnsCacheConfig>(
+              config, refresh_interval_.count(), random)),
       host_ttl_(PROTOBUF_GET_MS_OR_DEFAULT(config, host_ttl, 300000)),
       max_hosts_(PROTOBUF_GET_WRAPPED_OR_DEFAULT(config, max_hosts, 1024)) {
   tls_slot_->set([](Event::Dispatcher&) { return std::make_shared<ThreadLocalHostInfo>(); });
@@ -197,9 +202,13 @@ void DnsCacheImpl::finishResolve(const std::string& host,
   // Kick off the refresh timer.
   // TODO(mattklein123): Consider jitter here. It may not be necessary since the initial host
   // is populated dynamically.
-  // TODO(junr03): more aggressive refresh interval when DNS resolution fails.
-  // related issue: https://github.com/lyft/envoy-mobile/issues/673
-  primary_host_info.refresh_timer_->enableTimer(refresh_interval_);
+  if (status == Network::DnsResolver::ResolutionStatus::Success) {
+    failure_backoff_strategy_->reset();
+    primary_host_info.refresh_timer_->enableTimer(refresh_interval_);
+  } else {
+    primary_host_info.refresh_timer_->enableTimer(
+        std::chrono::milliseconds(failure_backoff_strategy_->nextBackOffMs()));
+  }
 }
 
 void DnsCacheImpl::runAddUpdateCallbacks(const std::string& host,